# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: UGC Net App API (Draft)
  version: '0.1.0'
  description: |
    Compact OpenAPI skeleton covering the main endpoints referenced in the pages docs. Use this
    as a starting point to generate server stubs, client SDKs, or API docs.
servers:
  - url: /api/v1
paths:
  /auth/login:
    post:
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
  /questions:
    post:
      summary: Create question
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionCreate'
      responses:
        '201':
          description: Created
  /questions/{id}:
    get:
      summary: Get question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Question
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
...
components:
  schemas:
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        display_name:
          type: string
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        display_name:
          type: string
...
    Question:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        answer_type:
          type: string
    QuestionCreate:
      type: object
...